change to require `{}` delimiters
we need this to ensure that judgements can be parsed without the
string parser consuming the whole input.

i chose { } to visually distringuish between strings and first
order/sexp terms.

there's a conversation to be had about whether having both `{}`
and () in the grammar is necessary. maybe `{}` is enough for the
inner ones too. for now, {} is forbidden anywhere except on the edges
of the string